feat：<数值系统>接入NumericComponent组件，规范化数值改变流程

+++ Server/Model/NKGMOBA/Battle/SkillSystem/Numeric/NumericChangeEvent_NotifyWatcher.cs
@@ -2,9 +2,9 @@
 {
 	// 分发数值监听
 	[Event(EventIdType.NumbericChange)]
-	public class NumericChangeEvent_NotifyWatcher: AEvent<long, NumericType, int>
+	public class NumericChangeEvent_NotifyWatcher: AEvent<long, NumericType, float>
 	{
-		public override void Run(long id, NumericType numericType, int value)
+		public override void Run(long id, NumericType numericType, float value)
 		{
 			Game.Scene.GetComponent<NumericWatcherComponent>().Run(numericType, id, value);
 		}
